Brealing: Remove Dialog.js
